backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator>=2.0.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Document Processing
pypdf2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
python-pptx==0.6.23
pillow==10.2.0
pytesseract==0.3.10

# LLM & AI
langchain>=0.1.4
langchain-openai>=0.0.5
langchain-google-genai>=0.0.6
langchain-community>=0.0.14
openai>=1.10.0
google-generativeai>=0.3.2
chromadb>=0.4.22
sentence-transformers>=2.3.1

# NLP
spacy==3.7.2
keybert==0.8.3

# Utilities
aiofiles==23.2.1
httpx==0.26.0
celery==5.3.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
